LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-target/rel-button-target-expected.txt


Harness Error (TIMEOUT), message = null

PASS <form rel=""> with <button formtarget>
FAIL <form rel="noopener"> with <button formtarget> assert_false: opener expected false got true
FAIL <form rel="noreferrer"> with <button formtarget> assert_equals: referrer expected "" but got "http://localhost:8800/html/semantics/forms/form-submission-target/rel-button-target.html"
PASS <form rel="opener"> with <button formtarget>
FAIL <form rel="noopener noreferrer"> with <button formtarget> assert_equals: referrer expected "" but got "http://localhost:8800/html/semantics/forms/form-submission-target/rel-button-target.html"
FAIL <form rel="noreferrer opener"> with <button formtarget> assert_equals: referrer expected "" but got "http://localhost:8800/html/semantics/forms/form-submission-target/rel-button-target.html"
FAIL <form rel="opener noopener"> with <button formtarget> assert_false: opener expected false got true
